# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebBanSachOnline API",
    "version": "1.0.0",
    "description": "Tài liệu API cho dự án WebBanSachOnline"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/books": {
      "get": {
        "summary": "Lấy danh sách sách",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchBy",
            "in": "query",
            "description": "Trường tìm kiếm: all | title | author | publisher | category_id | publish_year",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "Lọc chính xác theo mã thể loại (VD: VH, KT, TL)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sắp xếp: field-asc hoặc field-desc (VD: price-asc, created_at-desc)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công"
          }
        }
      }
    },
    "/books/latest": {
      "get": {
        "summary": "Lấy danh sách sách mới nhất",
        "tags": ["Books"],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": { "200": { "description": "Thành công" } }
      }
    },
    "/books/featured": {
      "get": {
        "summary": "Lấy sách đang giảm giá (is_on_sale = true)",
        "tags": ["Books"],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 12 } }
        ],
        "responses": { "200": { "description": "Trả về { items } danh sách sách đang sale" } }
      }
    },
    "/books/bestsellers": {
      "get": {
        "summary": "Lấy sách bán chạy nhất (dựa trên order_items)",
        "tags": ["Books"],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": { "200": { "description": "Trả về { items } danh sách sách bán chạy, sắp xếp theo số lượng đã bán" } }
      }
    },
    "/books/{bookId}": {
      "get": {
        "summary": "Xem chi tiết một cuốn sách",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "404": {
            "description": "Không tìm thấy"
          }
        }
      }
    },
    "/admin/books": {
      "post": {
        "summary": "Thêm sách mới (Admin)",
        "tags": [
          "Admin Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "book_id",
                  "title",
                  "author",
                  "category_id",
                  "price",
                  "publish_year",
                  "quantity"
                ],
                "properties": {
                  "book_id": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "author": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "quantity": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đã tạo"
          },
          "403": {
            "description": "Không có quyền"
          }
        }
      }
    },
    "/admin/books/{bookId}": {
      "patch": {
        "summary": "Cập nhật sách (Admin)",
        "tags": [
          "Admin Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã cập nhật"
          }
        }
      },
      "delete": {
        "summary": "Xóa sách (Admin)",
        "tags": [
          "Admin Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Đã xóa"
          }
        }
      }
    },
...
  }
}